separate utils from tools
