--- tests/legacy/integration/deps/wazuh_testing/tools/file.py
@@ -1,94 +1,0 @@
-import os
-from typing import Union, List
-import chardet
-import yaml
-
-
-def write_file(file_path: str, data: Union[List[str], str] = ''):
-    """
-    Write the specified data to the specified file.
-
-    Args:
-        file_path (str): The path to the file to write to.
-        data (List[str], str): The data to write to the file. This can either
-                               be a string or a list of strings.
-
-    Returns:
-        None
-    """
-    with open(file_path, 'w') as f:
-        f.writelines(data)
-
-
-def read_file_lines(path: str) -> List[str]:
-    """
-    Read the lines of a file and return them as a list of strings.
-
-    Args:
-        path (str): The path to the file to read.
-
-    Returns:
-        List[str]: A list of strings, each containing a line from the file.
-    """
-    with open(path, "r+") as f:
-        lines = f.readlines()
-    return lines
-
-
-def read_yaml(file_path):
-    """Read a YAML file from a given path, return a dictionary with the YAML data
-
-    Args:
-        file_path (str): Path of the YAML file to be readed
-
-    Returns:
-       dict: Yaml structure.
-    """
-    with open(file_path) as f:
-        return yaml.safe_load(f)
-
-
-def truncate_file(file_path: str) -> None:
-    """
-    Truncates the specified file.
-
-    Args:
-        file_path (str): The path to the file to be truncated.
-
-    Returns:
-        None
-    """
-    with open(file_path, "w") as f:
-        f.truncate()
-
-
-def get_file_encoding(file_path):
-    """Detect and return the file encoding.
-
-    Args:
-        file_path (str): File path to check.
-
-    Returns:
-        str: File encoding.
-
-    Raises:
-        ValueError: If could not find the file_path or is not a file.
-        TypeError: If could not detect the file encoding.
-    """
-    if not os.path.exists(file_path) or not os.path.isfile(file_path):
-        raise ValueError(f"{file_path} was not found or is not a file.")
-
-    # Read the file as bytes
-    with open(file_path, 'rb') as _file:
-        data = _file.read()
-
-    # Detect the content encoding
-    encoding = chardet.detect(data)['encoding']
-
-    if len(data) == 0:
-        return 'utf-8'
-
-    if encoding is None:
-        raise TypeError(f"Could not detect the {file_path} encoding")
-
-    return encoding
