+++ tests/legacy/integration/tests/test_analysisd/test_signature_id/test_valid_signature_id.py
@@ -46,7 +46,7 @@ from pathlib import Path
 
 from wazuh_testing.constants.paths.logs import OSSEC_LOG_PATH
 from wazuh_testing.modules.analysisd import patterns
-from wazuh_testing.tools.config import load_configuration_template, get_test_cases_data
+from wazuh_testing.utils.config import load_configuration_template, get_test_cases_data
 from wazuh_testing.tools.file_monitor import FileMonitor, generate_callback
 
 from . import CONFIGS_PATH, TEST_CASES_PATH, RULES_SAMPLE_PATH
